Deployment/requirements.txt
streamlit
pandas
numpy
scikit-learn == 1.0.2
tensorflow-cpu
pillow
opencv-python-headless==4.2.0.32